requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
plotly==5.17.0
matplotlib==3.8.1
seaborn==0.13.0
openai==1.3.0
langchain==0.0.350
pycaret==3.0.4
pandas-profiling==3.6.6
python-pptx==0.6.21
reportlab==4.0.4
openpyxl==3.1.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6